+++ Tests/RequestProcessor/Step/Common/SerializeResponseDataStepTest.php
@@ -3,6 +3,7 @@
 namespace Kami\ApiCoreBundle\Tests\RequestProcessor\Step\Common;
 
 use JMS\Serializer\Serializer;
+use Kami\ApiCoreBundle\RequestProcessor\Step\Common\BuildSerializerStep;
 use Kami\ApiCoreBundle\RequestProcessor\Step\Common\SerializeResponseDataStep;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
@@ -13,35 +14,32 @@ class SerializeResponseDataStepTest extends TestCase
 
     public function testCanBeConstructed()
     {
-        $serializerMock = $this->createMock(Serializer::class);
-
-        $step = new SerializeResponseDataStep($serializerMock);
+        $step = new SerializeResponseDataStep();
         $this->assertInstanceOf(SerializeResponseDataStep::class, $step);
     }
 
     public function testRequiresBefore()
     {
-        $serializerMock = $this->createMock(Serializer::class);
-
-        $step = new SerializeResponseDataStep($serializerMock);
-        $this->assertEquals([], $step->requiresBefore());
+        $step = new SerializeResponseDataStep();
+        $this->assertEquals([BuildSerializerStep::class], $step->requiresBefore());
     }
 
     public function testExecute()
     {
         $serializerMock = $this->createMock(Serializer::class);
-        $serializerMock->expects($this->any())->method('serialize')->willReturn($serializerMock);
+        $serializerMock->expects($this->any())->method('serialize')->willReturn('[]');
 
-        $step = new SerializeResponseDataStep($serializerMock);
+        $step = new SerializeResponseDataStep();
         $request = new Request();
         $request->attributes->set('_format', 'json');
         $step->setRequest($request);
         $step->setPreviousResponse(new ProcessorResponse($request, [
-            'response_data' => 1
+            'serializer' => $serializerMock,
+            'response_data' => []
         ]));
 
         $response = $step->execute();
         $this->assertInstanceOf(ProcessorResponse::class, $response);
-        $this->assertInstanceOf(Serializer::class, $response->getData()['response_data']);
+        $this->assertEquals('[]', $response->getData()['response_data']);
     }
 }