+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # CHANGE LOG
+## v0.2.1
+_28th, May 2018_
+
+* Refactored `BuildSelectQueryStep`
+* Moved exclusion logic to serialization
+* Added pagerfanta
+* Fixed minor (critical though) issues  
 
 ## v0.2.0
 _30th, April 2018_
 
 * Added `RequestProcessor\RequestProcessor`
 * Refactored whole workflow to use `Strategy` pattern
-* `Manager\ApiManager` to be removed
+* `Manager\ApiManager` was removed
 * Refactored `Security\AccessManager`
 * Added `Model\UserAwareInterface`
 * Improved testability